% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahChainFiles}
\alias{ahChainFiles}
\title{ahChainFile}
\format{
A data frame with 1113 rows and 2 columns
\describe{
    \item{id}{is the AnnotationHub id}
    \item{chainFile}{is the *.chain.gz file}
}
}
\usage{
data(ahChainFiles)
}
\description{
A data frame containing the AnnotationHub id related to the *.chain.gz file,
specific for each species and genome. The file name reflects the assembly
conversion data contained within in the format <db1>To<Db2>.over.chain.gz.
For example, a file named mm10ToHg19.over.chain.gz file contains the
liftOver data needed to convert mm10 (Mouse GRCm38) coordinates to hg19
(Human GRCh37).
}
\examples{
data(ahChainFiles)

}
\keyword{datasets}
