% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahRepeatMasker}
\alias{ahRepeatMasker}
\title{ahRepeatMasker}
\format{
A data frame with 84 rows and 3 columns
\describe{
    \item{id}{is the AnnotationHub id}
    \item{species}{is the species}
    \item{genome}{is the genome assembly}
}
}
\usage{
data(ahRepeatMasker)
}
\description{
A data frame containing the AnnotationHub id related to the repeatMasker db,
specific for each species and genome.
}
\examples{
data(ahRepeatMasker)

}
\keyword{datasets}
