% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateSNPsGWAS.R
\name{annotateSNPsGWAS}
\alias{annotateSNPsGWAS}
\title{Annotate GWAS SNPs}
\usage{
annotateSNPsGWAS(targets, genome = "hg19", pathToTraits = NULL)
}
\arguments{
\item{targets}{A list containing the target regions to analyze.
It can be generated with \code{\link{getSeqsFromGRs}}.}

\item{genome}{A string specifying the human genome assembly the target
regions came from. Possible options are hg38 or hg19.
Current image for GWAS SNPS coordinates is hg38. If hg19 is specified
SNPs coordinates are realtime liftOver to hg19 coordinates.
Internally at first the function \code{\link[gwascat]{makeCurrentGwascat}}
from the \code{\link{gwascat}} package is used to get the more recent image.
If an error occurs due to connection to the db, data(ebicat37) or
data(ebicat38) are used. Default value is "hg19".}

\item{pathToTraits}{A string containing the path to the traits.txt
file. contains diseases/traits specified by the user. It must
have one column with header id. By default pathToTraits is set to NULL and
the file it is searched in the working directory. If traits.txt is located
in a different directory then the path needs to be specified. If this file is
absent or empty SNPs associated with all diseases/traits in
the GWAS catalog are considered in the SNPs analysis.}
}
\value{
A list.
}
\description{
The function annotateSNPsGWAS() annotates GWAS SNPs located
in the region flanking the back-spliced junctions of each circRNA.
SNPs information including the corresponding genomic coordinates are
retrived from the GWAs catalog database.
The user can restric the analysis to specific traits/diseases. These must go
in the file traits.txt. If this file is absent or empty, all traits in the
GWAS catalog are considered in the analysis.
}
\examples{
# Load data frame containing detected back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Annotate the first 10 back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1:10, ], gtf,
isRandom = FALSE)

# Retrieve targets
targets <- getSeqsFromGRs(
    annotatedBSJs,
    lIntron = 200,
    lExon = 10,
    type = "ie",
    species = "Hsapiens",
    genome = "hg19")

# Annotate GWAS SNPs
snpsGWAS <- annotateSNPsGWAS(targets, genome = "hg19")

}
