% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attractSpecies}
\alias{attractSpecies}
\title{attractSpecies}
\format{
A data frame with 37 rows and 1 columns
\describe{
    \item{species}{is the species}
}
}
\usage{
data(attractSpecies)
}
\description{
A data frame containing the species for which RBP motifs are available
in ATtRACT db. See also \url{http://attract.cnic.es}.
}
\examples{
data(attractSpecies)

}
\keyword{datasets}
