% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{checkBSJsDF}
\alias{checkBSJsDF}
\title{Check backSplicedJunctions data frame}
\usage{
checkBSJsDF(backSplicedJunctions, addColNames = NULL)
}
\arguments{
\item{backSplicedJunctions}{A data frame containing back-spliced junction
coordinates.}

\item{addColNames}{A string vector containing the columns that the data frame
must contains in addition to the basic columns. By default addColNames is set
to NULL.}
}
\value{
A data frame
}
\description{
The function checkBSJsDF() verifies that the functions to import
the detected circRNAs generate the correct data structure and content.
}
\examples{
# Load short version of the gencode v19 annotation file
data("gtf")

# Path to an example file containing circRNA detected by MapSplice2
pathToFile <- system.file("extdata", "mapsplice/circRNAs_001.txt",
    package="circRNAprofiler")

# Inner function.
# Import circRNAs.
backSplicedJunctions <- importMapSplice(pathToFile, gtf)

# Inner function
# check table
checkBSJsDF(backSplicedJunctions, addColNames = "coverage")

}
