% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMirSites.R
\name{compareSequences}
\alias{compareSequences}
\title{Align sequences}
\usage{
compareSequences(seq1, seq2, isGUMatch = TRUE)
}
\arguments{
\item{seq1}{A character string.}

\item{seq2}{A character string.}

\item{isGUMatch}{A logical specifying whether to consider G:U pair as a match
or as a mismatch. Default value is TRUE.}
}
\value{
A character string.
}
\description{
The function compareSequences() combines two equal-length
strings that are assumed to be aligned into a single character string.
The Watson-Crick matches are replaced with "w", the mismatches with "m",
G:U matches are replaced either with "n" or with "w" based on the value of
isGUMatch argument.
}
\examples{
# Inner function.
compareSequences("UGGCGUU", "AGGUUUG", isGUMatch = TRUE)

}
