% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifs.R
\name{computeMotifs}
\alias{computeMotifs}
\title{Compute and search motifs of a given length in the target sequences.}
\usage{
computeMotifs(targets, width = 6)
}
\arguments{
\item{targets}{A list containing the target sequences to analyze.
It can be generated with: \code{\link{getCircSeqs}} or
\code{\link{getSeqsAcrossBSJs}} or \code{\link{getSeqsFromGRs}}.}

\item{width}{An integer specifying the length of all possible motifs to
extract from the target sequences. Default value is 6.}
}
\value{
A data frame.
}
\description{
The function computeMotifs() first computes all possible motifs
of a given length and subsequently searches those motifs in the target
sequences. Only motifs present at least 1 time are kept.
}
\examples{
# Load data frame containing detected back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Example with the first back-spliced junctions.
# Multiple back-spliced junctions can also be analyzed at the same time.

# Annotate detected back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1, ], gtf)

# Retrieve targets
targets <- getSeqsFromGRs(
    annotatedBSJs,
    lIntron = 200,
    lExon = 10,
    type = "ie",
    species = "Hsapiens",
    genome = "hg19")

# Inner function
# Compute motifs
computedMotifs <- computeMotifs(
    targets,
    width = 6)


}
\keyword{internal}
