% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMirSites.R
\name{createMiRsitesList}
\alias{createMiRsitesList}
\title{Create list to store miR results}
\usage{
createMiRsitesList(targets, microRNAs)
}
\arguments{
\item{targets}{A list containing the target sequences to analyze.
It can be generated with \code{\link{getCircSeqs}}.}

\item{microRNAs}{A data frame containing microRNA sequences to analyze.
It can be generated with \code{\link{getMiRseqs}}.}
}
\value{
A list.
}
\description{
Create list to store miR results
}
\examples{
# Load data frame containing predicted back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Annotate the first 3 back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1:3, ], gtf,
isRandom = FALSE)

# Retrieve target sequences
targets <- getCircSeqs(
    annotatedBSJs,
    gtf,
    species = "Hsapiens",
    genome = "hg19")

# Inner functions
# Retrieve microRNA sequences
microRNAs <- getMiRseqs(miRBaseLatestRelease = TRUE,
    miRspeciesCode = "hsa")

# Create list
miRsites <- createMiRsitesList(targets, microRNAs)


}
