% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifs.R
\name{createMotifsList}
\alias{createMotifsList}
\title{Create list to store motif results}
\usage{
createMotifsList(targets, filteredMotifs)
}
\arguments{
\item{targets}{A list containing the target sequences to analyze.
It can be generated with \code{\link{getCircSeqs}}.}

\item{filteredMotifs}{A data frame containing motifs to analyze.
It can be generated with \code{\link{filterMotifs}}.}
}
\value{
A list.
}
\description{
Create list to store motif results
}
\examples{
# Load data frame containing detected back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Example with the first back-spliced junctions.
# Multiple back-spliced junctions can also be analyzed at the same time.

# Annotate detected back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1, ], gtf)

# Retrieve targets
targets <-
getSeqsFromGRs(
    annotatedBSJs,
    lIntron = 200,
    lExon = 10,
    type = "ie",
    species = "Hsapiens",
    genome = "hg19")

# Inner function
# Compute motifs
computedMotifs <- computeMotifs(
    targets,
    width = 6)

# Inner function
# Filter motifs
filteredMotifs <- filterMotifs(
    computedMotifs,
    species = "Hsapiens",
    rbp = TRUE,
    reverse = FALSE)

# Inner function
# Create list
motifs <- createMotifsList(
    targets,
    filteredMotifs)


}
\keyword{internal}
