% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifs.R
\name{filterMotifs}
\alias{filterMotifs}
\title{Filter motifs}
\usage{
filterMotifs(computedMotifs, species = "Hsapiens", rbp = TRUE,
  reverse = FALSE, pathToMotifs = NULL)
}
\arguments{
\item{computedMotifs}{A character vector containing motifs extracted from
the target sequences. It can be generated with \code{\link{computeMotifs}}.}

\item{species}{A string specifying the species of the RBP motifs to use.
Type data(attractSpecies) to see the possible options.
Default values is "Hsapiens".}

\item{rbp}{A logical specifying whether to report only motifs matching
with known RBP motifs from ATtRACT database or user motifs specified in
motifs.txt. If FALSE is specified only motifs that do not match with any of
these motifs are reported. Default value is TRUE.}

\item{reverse}{A logical specifying whether to reverse the motifs collected
from ATtRACT database and from motifs.txt. If TRUE is specified all the
motifs are reversed and analyzed together with the direct motifs as they are
reported in the ATtRACT db and motifs.txt. Default value is FALSE.}

\item{pathToMotifs}{A string containing the path to the motifs.txt
file. The file motifs.txt contains motifs/regular expressions specified
by the user. It must have 3 columns with headers:
- id (1st column): name of the motif. - e.g. RBM20 or motif1.
- motif (2nd column): motif/pattern to search.
- length (3rd column): length of the motif.
By default pathToMotifs is set to NULL and the file it is searched in the
working directory. If motifs.txt is located in a different directory then
the path needs to be specified. If this file is absent or empty only the
motifs of RNA Binding Proteins in the ATtRACT database are considered in
the motifs analysis.}
}
\value{
A data frame
}
\description{
The function filterMotifs() finds motifs that match with motifs
of known RNA Binding Proteins (RBPs) deposited in the ATtRACT database and
with motifs specified by the user reported in motifs.txt. Subsequently,
they are filtered based on the value of rbp argument.
}
\examples{
# Load data frame containing detected back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Example with the first back-spliced junctions.
# Multiple back-spliced junctions can also be analyzed at the same time.

# Annotate detected back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1, ], gtf)

# Retrieve targets
targets <- getSeqsFromGRs(
    annotatedBSJs,
    lIntron = 200,
    lExon = 10,
    type = "ie",
    species = "Hsapiens",
    genome = "hg19" )

# Inner function
# Compute motifs
computedMotifs <- computeMotifs(
    targets,
    width = 6)

# Inner function
# Filter motifs
filteredMotifs <- filterMotifs(
   computedMotifs,
   species = "Hsapiens",
   rbp = TRUE,
   reverse = FALSE)

}
\keyword{internal}
