% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGTF.R
\name{formatGTF}
\alias{formatGTF}
\title{Format annotation file}
\usage{
formatGTF(pathToGTF = NULL)
}
\arguments{
\item{pathToGTF}{A string containing the path to the GTF file.
Use the same annotation file used during the RNA-seq mapping procedure.
If .gtf file is not present in the current working directory the full path
should be specified.}
}
\value{
A data frame.
}
\description{
The function formatGTF() formats the given annotation file.
}
\examples{
gtf <- formatGTF()

}
