% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateBSJs.R
\name{getAllTranscripts}
\alias{getAllTranscripts}
\title{Retrieve transcripts}
\usage{
getAllTranscripts(gtf, gene, exJ1, exJ2)
}
\arguments{
\item{gtf}{A data frame containing genome annotation information,
generated with \code{\link{formatGTF}}.}

\item{gene}{A string specifying the gene name whose exon coordinates
overlap that of the given back-spliced junctions.}

\item{exJ1}{A string specifying one of the two back-spliced junction
coordinates.}

\item{exJ2}{A string specifying one of the two back-spliced junction
coordinates.}
}
\value{
A data frame.
}
\description{
The function getAllTranscripts() retrieves all transcripts whose
exon coordinates overlap that of the given back-spliced junctions.
}
\examples{
# Inner function.

# Load short version of the gencode v19 annotation file
data("gtf")

allTranscripts <- getAllTranscripts(gtf, "SYCP2", "58497514", "58497445")

}
