% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateBSJs.R
\name{getAnnotatedBSJsColNames}
\alias{getAnnotatedBSJsColNames}
\title{Return column names}
\usage{
getAnnotatedBSJsColNames()
}
\value{
A character vector.
}
\description{
The function getAnnotatedBSJsColNames() returns the column
names.
}
\examples{
# Inner function.
getAnnotatedBSJsColNames()

}
\keyword{internal}
