% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getBasicColNames}
\alias{getBasicColNames}
\title{Return column names}
\usage{
getBasicColNames()
}
\value{
A character vector
}
\description{
The function getBasicColNames() returns the basic column names.
}
\examples{
# Inner function
getBasicColNames()

}
\keyword{internal}
