% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqsFromGRs.R
\name{getBSEcoords}
\alias{getBSEcoords}
\title{Retrieve back-spliced exon coordinates}
\usage{
getBSEcoords(annotatedBSJs)
}
\arguments{
\item{annotatedBSJs}{A data frame with the annotated back-spliced junctions.
This data frame can be generated with \code{\link{annotateBSJs}}.}
}
\value{
A data frame.
}
\description{
The function getBSEcoords() retrieves the coordinates of the
back-spliced exons from the annotateBSJs data frame.
}
\examples{
# Load data frame containing predicted back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Example with the first back-spliced junction
# Multiple back-spliced junctions can also be analyzed at the same time.

# Annotate predicted back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1, ], gtf)

# Inner function
# Retrieve back-spliced exon coordinates
getBSEcoords(annotatedBSJs)


}
