% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBackSplicedJunctions.R
\name{getDetectionTools}
\alias{getDetectionTools}
\title{Create data frame with circRNA detection codes}
\usage{
getDetectionTools()
}
\value{
A data frame
}
\description{
The function getDetectionTools() creates a data frame
containing the codes corresponding to each circRNA detection tool for which
a specific import function has been developed.
}
\examples{
getDetectionTools()

}
