% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateBSJs.R
\name{getFlankIntronFirst}
\alias{getFlankIntronFirst}
\title{Retrieve flanking intron coordinates}
\usage{
getFlankIntronFirst(transcriptToAnalyze, bsExons)
}
\arguments{
\item{transcriptToAnalyze}{A data frame with the genomic coordinates of all
exons present in the transcript whose exon coordinates overlap that
of the given back-spliced junctions.}

\item{bsExons}{A data frame with the back-spliced exon coordinates.}
}
\value{
A data frame.
}
\description{
The function getFlankIntronFirst() retrieves the genomic
coordinates of the introns flanking the back-spliced exons given in input.
This function is called when one of the two back-spliced exons is the
first of the transcript.
}
\examples{
# Inner function.

# Load short version of the gencode v19 annotation file
data("gtf")

transcriptToAnalyze <- gtf[gtf$transcript_id== "ENST00000357552.3",]
bsExons <- transcriptToAnalyze[c(1,2),]

intronCoords <- getFlankIntronFirst(transcriptToAnalyze, bsExons)

}
