% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqsFromGRs.R
\name{getGRcolNames}
\alias{getGRcolNames}
\title{The function getGRcolNames() returns the column names}
\usage{
getGRcolNames()
}
\value{
A character vector.
}
\description{
The function getGRcolNames() returns the column names
}
\examples{
# Inner function
getGRcolNames()

}
\keyword{internal}
