% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqsFromGRs.R
\name{getIECoords}
\alias{getIECoords}
\title{Retrive coordinates of genomic ranges surrounding back-spliced
junction coordinates}
\usage{
getIECoords(annotatedBSJs, lIntron, lExon)
}
\arguments{
\item{annotatedBSJs}{A data frame with the annotated back-spliced junctions.
This data frame can be generated with \code{\link{annotateBSJs}}.}

\item{lIntron}{An integer indicating how many nucleotides are taken from
the introns flanking the back-spliced junctions. This number must be
positive.}

\item{lExon}{An integer indicating how many nucleotides are taken from
the back-spliced exons starting from the back-spliced junctions. This number
must be positive.}
}
\value{
A data frame.
}
\description{
The function getIECoords() retrieves the coordinates of the
genomic ranges defined by the values of the arguments lIntron and lExon.
The start point are the back-spliced junction coordinaes reported in the
annotateBSJs data frame.
}
\examples{
# Load data frame containing predicted back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Example with the first back-spliced junction.
# Multiple back-spliced junctions can also be analyzed at the same time.

# Annotate predicted back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1, ], gtf)

# Inner function
# Retrieve genomic range coordinates
getIECoords(annotatedBSJs, lIntron = 101, lExon = 10)


}
\keyword{internal}
