% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLength.R
\name{getLength}
\alias{getLength}
\title{Retieve feature length}
\usage{
getLength(annotatedBSJs)
}
\arguments{
\item{annotatedBSJs}{A data frame with the annotated back-spliced junctions.}
}
\value{
A data frame.
}
\description{
The function getLength() calculates the length (nt) of the
back-spliced exons and the corresponding flanking introns.
}
\examples{
# Inner function
annotatedBSJs <- data.frame(
    matrix(nrow = 1, ncol = length(getAnnotatedBSJsColNames())))

colnames(annotatedBSJs) <- getAnnotatedBSJsColNames()
annotatedBSJs$id <- "SYCP2:-:chr20:58497514:58497445"
annotatedBSJs$startUpIntron <- 58507116
annotatedBSJs$endUpIntron <- 58497515
annotatedBSJs$startUpBSE <- 58497514
annotatedBSJs$endUpBSE <- 58497445
annotatedBSJs$startDownBSE <- 58497514
annotatedBSJs$endDownBSE <- 58497445
annotatedBSJs$startDownIntron <- 58497444
annotatedBSJs$endDownIntron <- 58496509

# Retrieve length
getLength(annotatedBSJs)

}
\keyword{internal}
