% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMirSites.R
\name{getMatches}
\alias{getMatches}
\title{Retrieve sequence matches}
\usage{
getMatches(seq1, seq2, isSeed = TRUE)
}
\arguments{
\item{seq1}{A character string from the circRNA sequence.}

\item{seq2}{A character string from the miR sequence.}

\item{isSeed}{A logical indicating whether the analysis is perfomed on the
seed region. Default value is TRUE.}
}
\value{
A dataframe.
}
\description{
The function getMatches() retrives the number of total matches,
continuous canonical Watson and Crick matches and a non-canonical matches
(G:U) between 2 given sequences of the same length. The analysis is
performed by comparing the 2 sequences without gap and by introducing a
gap in each position of one of the 2 sequences at the time.
}
\examples{
# Inner function.
matches <- getMatches("AGGCGUU", "AGGCGUU", isSeed = TRUE)


}
