% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMirSites.R
\name{getMiRseqs}
\alias{getMiRseqs}
\title{Retrive miRNA sequences}
\usage{
getMiRseqs(miRBaseLatestRelease = TRUE, miRspeciesCode = "hsa",
  pathToMiRs = NULL)
}
\arguments{
\item{miRBaseLatestRelease}{A logical specifying whether to download the
latest release of the mature sequences of the microRNAs from miRBase
(\url{http://www.mirbase.org/ftp.shtml}). If TRUE is specifed then the
latest release is automatically downloaded. If FALSE is specified then a
file named mature.fa containing fasta format sequences of all mature miRNA
sequences previously downloaded by the user from mirBase must be present
in the working directory. Default value is TRUE.}

\item{miRspeciesCode}{A string specifying the species code (3 letters) as
reported in miRBase. E.g. to analyze the mouse microRNAs specify "mmu",
to analyze the human microRNAs specify "hsa". Type data(miRspeciesCode)
to see the available codes and the corresponding species reported
in miRBase 22 release. Default value is "hsa".}

\item{pathToMiRs}{A string containing the path to the miRs.txt file.
The file miRs.txt contains the microRNA ids from miRBase
specified by the user. It must have one column with header id. The first row
must contain the miR name starting with the ">", e.g >hsa-miR-1-3p. The
sequences of the miRs will be automatically retrieved from the mirBase latest
release or from the given mature.fa file, that should be present in the
working directory. By default pathToMiRs is set to NULL and the file it is
searched in the working directory. If miRs.txt is located in a different
directory then the path needs to be specified. If this file is absent or
empty, all miRs of the species are retrieved.}
}
\value{
A data frame.
}
\description{
The function getMiRseqs() retrieves miRNA sequences from
miRbase data base
}
\examples{
# Inner function
microRNAs <- getMiRseqs(miRBaseLatestRelease = TRUE,
    miRspeciesCode = "hsa")

}
\keyword{internal}
