% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegexPattern.R
\name{getRegexPattern}
\alias{getRegexPattern}
\title{Convert IUPAC sequence to an regular expression}
\usage{
getRegexPattern(iupacSeq, isDNA = FALSE)
}
\arguments{
\item{iupacSeq}{A character string with IUPAC codes.}

\item{isDNA}{A logical specifying whether the iupacSeq is DNA or RNA.
Deafult value is FALSE.}
}
\value{
A character string.
}
\description{
The function getRegexPattern() converts a nucleotide sequence
with IUPAC codes to an regular expression.
}
\examples{
regextPattern <- getRegexPattern("CGUKMBVNN", isDNA = FALSE)

}
