% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateRepeats.R
\name{getRepeatsColNames}
\alias{getRepeatsColNames}
\title{Return column names}
\usage{
getRepeatsColNames()
}
\value{
A character vector.
}
\description{
The function getRepeatsColNames() returns the column names.
}
\examples{
# Inner function
getRepeatsColNames()

}
