% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqsAcrossBSJs.R
\name{getSeqsAcrossBSJs}
\alias{getSeqsAcrossBSJs}
\title{Retrieve back-spliced junction sequences}
\usage{
getSeqsAcrossBSJs(annotatedBSJs, gtf, species = "Hsapiens",
  genome = "hg19")
}
\arguments{
\item{annotatedBSJs}{A data frame with the annotated back-spliced junctions.
It can be generated with \code{\link{annotateBSJs}}.}

\item{gtf}{A dataframe containing genome annotation information. It can be
generated with \code{\link{formatGTF}}.}

\item{species}{A string specifying the species from which to retrieve the
sequences. For Example: Mmusculus or Hsapiens.
See available.genomes() from BSgenome package to see all the species.
Default value is "Hsapiens".}

\item{genome}{A string specifying the genome assembly from which to retrieve
the sequences. For Example: mm10 or mm9 for Mouse, hg38 or hg19 for Human.
See available.genomes() from BSgenome package to see all genomes.
Default value is "hg19".}
}
\value{
A list.
}
\description{
The function getSeqsAcrossBSJs() retrieves
the sequences across the back-spliced junctions. A total of 11 nucleotides
from each side of the back-spliced junction are taken and concatenated
together.
}
\examples{
# Load a data frame containing detected back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Annotate the first 10 back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1:10, ], gtf)

# Retrieve target sequences
targets <- getSeqsAcrossBSJs(
    annotatedBSJs,
    gtf,
    species = "Hsapiens",
    genome = "hg19")


}
