% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqsFromGRs.R
\name{getSeqsFromGRs}
\alias{getSeqsFromGRs}
\title{Retrieve sequences flanking back-spliced junctions}
\usage{
getSeqsFromGRs(annotatedBSJs, lIntron = 100, lExon = 10, type = "ie",
  species = "Hsapiens", genome = "hg19")
}
\arguments{
\item{annotatedBSJs}{A data frame with the annotated back-spliced junctions.
This data frame can be generated with \code{\link{annotateBSJs}}.}

\item{lIntron}{An integer indicating how many nucleotides are taken from
the introns flanking the back-spliced junctions. This number must be
positive. Default value is 100.}

\item{lExon}{An integer indicating how many nucleotides are taken from
the back-spliced exons starting from the back-spliced junctions. This number
must be positive. Default value is 10.}

\item{type}{A character string specifying the sequences to retrieve.
If type = "ie" the sequences are retrieved from the the genomic ranges
defined by using the lIntron and lExon given in input. If type = "bse"
the sequences of the back-spliced exons are retrieved. If type = "fi"
the sequences of the introns flanking the back-spliced exons are retrieved.
Default value is "ie".}

\item{species}{A string specifying the species from which to retrieve the
sequences. For Example: Mmusculus or Hsapiens.
See available.genomes() from BSgenome package to see all the species.
Default value is "Hsapiens".}

\item{genome}{A string specifying the genome assembly from which to retrieve
the sequences. For Example: mm10 or mm9 for Mouse, hg38 or hg19 for Human.
See available.genomes() from BSgenome package to see all genomes.
Default value is "hg19".}
}
\value{
A list.
}
\description{
The function getSeqsFromGRs() includes 3 modules to retrieve
3 types of sequences. Sequences of the introns flanking back-spliced
junctions, sequences from a defined genomic window surrounding the
back-spliced junctions and sequences of the back-spliced exons.
}
\examples{
# Load data frame containing predicted back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Annotate the first 10 back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1:10, ], gtf)

# Retrieve target sequences
targets <- getSeqsFromGRs(
    annotatedBSJs,
    lIntron = 200,
    lExon = 10,
    type = "ie",
    species = "Hsapiens",
    genome = "hg19")


}
