% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getTargetsColNames}
\alias{getTargetsColNames}
\title{Return column names}
\usage{
getTargetsColNames()
}
\value{
A character vector.
}
\description{
The function getTargetsColNames() returns the column names.
}
\examples{
# Inner function
getTargetsColNames()

}
