% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifs.R
\name{getUserAttractMotifs}
\alias{getUserAttractMotifs}
\title{Retrieve ATtRACT motifs}
\usage{
getUserAttractMotifs(species = "Hsapiens", reverse = FALSE,
  pathToMotifs = NULL)
}
\arguments{
\item{species}{A string specifying the species of the RBP motifs to use.
Type data(attractSpecies) to see the possible options.
Default value is "Hsapiens".}

\item{reverse}{A logical specifying whether to reverse the motifs collected
from ATtRACT database and from motifs.txt. If TRUE is specified all the
motifs are reversed and analyzed together with the direct motifs as they are
reported in the ATtRACT db and motifs.txt. Default value is FALSE.}

\item{pathToMotifs}{A string containing the path to the motifs.txt
file. The file motifs.txt contains motifs/regular expressions specified
by the user. It must have 3 columns with headers:
- id (1st column): name of the motif. - e.g. RBM20 or motif1.
- motif (2nd column): motif/pattern to search.
- length (3rd column): length of the motif.
By default pathToMotifs is set to NULL and the file it is searched in the
working directory. If motifs.txt is located in a different directory then
the path needs to be specified. If this file is absent or empty only the
motifs of RNA Binding Proteins in the ATtRACT database are considered in
the motifs analysis.}
}
\value{
A data frame.
}
\description{
The function getUserAttractMotifs() reads the user motifs
in motifs.txt and retrieves the motifs deposited in the ATtRACT database
(\url{http://attract.cnic.es}).
}
\examples{
# Inner function
userAttractMotifs <- getUserAttractMotifs(species = "Hsapiens",
    reverse = FALSE)

}
