% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtf}
\alias{gtf}
\title{gtf}
\format{
An object of class \code{data.frame} with 4401 rows and 9 columns.
}
\usage{
data(gtf)
}
\description{
A data frame containing a short version of the already formatted gencode v19
based-genome annotations. This data frame can be generated with
\code{\link{formatGTF}}.
}
\examples{
data(gtf)

}
\keyword{datasets}
