% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwasTraits}
\alias{gwasTraits}
\title{gwasTraits}
\format{
A data frame with 653 rows and 1 columns
\describe{
    \item{id}{is the trait/disease as reported in the GWAS catalog}
}
}
\usage{
data(gwasTraits)
}
\description{
A data frame containing the traits/disease extracted on the 31th October 2018
from the GWAS catalog. See also \url{ https://www.ebi.ac.uk/gwas/} for more
detail.
}
\examples{
data(gwasTraits)

}
\keyword{datasets}
