% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initCircRNAprofiler.R
\name{initCircRNAprofiler}
\alias{initCircRNAprofiler}
\title{Initialize the project folder}
\usage{
initCircRNAprofiler(projectFolderName, detectionTools)
}
\arguments{
\item{projectFolderName}{A character string specifying the name of the
project folder.}

\item{detectionTools}{A character vector specifying the tools
used for circRNA detection. The following options are allowed: mapsplice,
nclscan, knife, circexplorer2, circmarker, uroborus and other. The user can
choose 1 or multiple tools. Subfolders named with the tools specified will
be generated under the project folder.}
}
\value{
A NULL object
}
\description{
The function initCircRNAprofiler() initializes the project
forlder.
}
\examples{
initCircRNAprofiler(projectFolderName = "circProject",
    detectionTools = "mapsplice")


}
