% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iupac}
\alias{iupac}
\title{iupac}
\format{
A data frame with 18 rows and 4 columns
\describe{
    \item{code}{is the IUPAC nucleotide code}
    \item{base}{is the base}
    \item{regexDNA}{is the regular expression of the corresponding
    IUPAC code}
    \item{regexRNA}{is the regular expression of the corresponding
    IUPAC code}
}
}
\usage{
data(iupac)
}
\description{
A data frame containing IUPAC codes and the correspoding regular expression.
}
\examples{
data(iupac)

}
\keyword{datasets}
