% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memeDB}
\alias{memeDB}
\title{memeDB}
\format{
A character vector with 25 rows ans 2 columns
\describe{
    \item{path}{is the file path }
    \item{index}{index of the file path }
}
}
\usage{
data(memeDB)
}
\description{
A dataframe containing the file paths of the files containing the RBP motifs
in meme format available in MEME db. See also \url{http://meme-suite.org/doc/download.html}.
File paths extracted from RNA folder of motif_databases.12.19.tgz
(Motif Databases updated 28 Oct 2019)
}
\examples{
data(memeDB)

}
\keyword{datasets}
