% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBackSplicedJunctions.R
\name{mergeBSJunctions}
\alias{mergeBSJunctions}
\title{Group circRNAs identified by multiple prediction tools}
\usage{
mergeBSJunctions(backSplicedJunctions, gtf)
}
\arguments{
\item{backSplicedJunctions}{A data frame containing back-spliced junction
coordinates and counts generated with \code{\link{getBackSplicedJunctions}}.}

\item{gtf}{A data frame containing genome annotation information,
generated with \code{\link{formatGTF}}.}
}
\value{
A data frame.
}
\description{
The function mergeBSJunctions() shrinks the data frame by
grouping back-spliced junctions commonly identified by multiple
detection tools. The read counts of the samples reported in the final
data frame will be the ones of the tool that detected the highest total mean.
All the tools that detected the back-spliced junctions are then listed
in the column tool of the final data frame.
See \code{\link{getDetectionTools}}  for more detail about the code
corresponding to each circRNA detection tool.
}
\examples{
# Load detected back-soliced junctions
data("backSplicedJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Merge commonly identified circRNAs
mergedBSJunctions <- mergeBSJunctions(backSplicedJunctions, gtf)


}
