% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRspeciesCodes}
\alias{miRspeciesCodes}
\title{miRspeciesCodes}
\format{
A data frame with 271 rows and 2 columns
\describe{
    \item{code}{is the unique identifier of the species as reported in
    miRBase db}
    \item{species}{is the species}

}
}
\usage{
data(miRspeciesCodes)
}
\description{
A data frame containing the code of the species as reported in miRBase
22 release. See also \url{http://www.mirbase.org}
}
\examples{
data(miRspeciesCodes)

}
\keyword{datasets}
