% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlots.R
\name{plotMotifs}
\alias{plotMotifs}
\title{Plot motifs analysis results}
\usage{
plotMotifs(mergedMotifsFTS, mergedMotifsBTS, log2FC = 1, nf1 = 1,
  nf2 = 1, df1Name = "foreground", df2Name = "background")
}
\arguments{
\item{mergedMotifsFTS}{A data frame containing the number of occurences
of each motif found in foreground target sequences (e.g from detected
back-spliced junctions). It can be generated with the
\code{\link{mergeMotifs}}.}

\item{mergedMotifsBTS}{A data frame containing the number of occurences
of each motif found in the background target sequences (e.g. from
random back-spliced junctions). It can be generated with the
\code{\link{mergeMotifs}}.}

\item{log2FC}{An integer specifying the log2FC cut-off. Default value is 1.}

\item{nf1}{An integer specifying the normalization factor for the
first data frame mergedMotifsFTS. The occurrences of each motif are divided
by nf1. The normalized values are then used for fold-change calculation.
Set this to the number of target sequences (e.g from detected
back-spliced junctions) where the motifs were extracted from.
Default value is 1.}

\item{nf2}{An integer specifying the normalization factor for the
second data frame mergedMotifsBTS. The occurrences of each motif are divided
by nf2. The normalized values are then used for fold-change calculation.
Set this to the number of target sequences (e.g from random
back-spliced junctions) where the motifs were extracted from.
Default value is 1.

NOTE: By setting nf1 and nf2 equals to 1 the number of target sequences
(e.g detected Vs randomly selected) where the motifs were extrated from,
is supposed to be the same.}

\item{df1Name}{A string specifying the name of the first data frame. This
will be displayed in the legend of the plot. Deafult value is "foreground".}

\item{df2Name}{A string specifying the name of the first data frame. This
will be displayed in the legend of the plot. Deafult value is "background".}
}
\value{
A ggplot object.
}
\description{
The function plotMotifs() generates 2 bar charts showing the
log2FC and the number of occurences of each motif found in the target
sequences (e.g detected Vs randomly selected).
}
\examples{
# Load data frame containing detected back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Annotate the first 10 back-spliced junctions
annotatedFBSJs <- annotateBSJs(mergedBSJunctions[1:10, ], gtf, isRandom = FALSE)

# Get random back-spliced junctions
randomBSJunctions <- getRandomBSJunctions(gtf, n = 10, f = 10)

# Annotate random back-spliced junctions
annotatedBBSJs <- annotateBSJs(randomBSJunctions, gtf, isRandom = TRUE)

# Retrieve target sequences from detected back-spliced junctions
targetsFTS <- getSeqsFromGRs(
   annotatedFBSJs,
   lIntron = 200,
   lExon = 10,
   type = "ie",
   species = "Hsapiens",
   genome = "hg19")

# Retrieve target sequences from random back-spliced junctions
targetsBTS <- getSeqsFromGRs(
   annotatedBBSJs,
   lIntron = 200,
   lExon = 10,
   type = "ie",
   species = "Hsapiens",
   genome = "hg19")


# Get motifs
motifsFTS <- getMotifs(
    targetsFTS,
    width = 6,
    species = "Hsapiens",
    rbp = TRUE,
    reverse = FALSE)

motifsBTS <- getMotifs(
    targetsBTS,
    width = 6,
    species = "Hsapiens",
    rbp = TRUE,
    reverse = FALSE)

# Merge motifs
mergedMotifsFTS <- mergeMotifs(motifsFTS)
mergedMotifsBTS <- mergeMotifs(motifsBTS)

# Plot
p <- plotMotifs(
    mergedMotifsFTS,
    mergedMotifsBTS,
    log2FC = 2,
    nf1 = nrow(annotatedFBSJs),
    nf2 = nrow(annotatedBBSJs),
    df1Name = "foreground",
    df2Name = "background")

}
