% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMirSites.R
\name{rearrangeMiRres}
\alias{rearrangeMiRres}
\title{Rearrange miR results}
\usage{
rearrangeMiRres(miRsites)
}
\arguments{
\item{miRsites}{A list containing the miR sites found in the RNA
target sequence. it can be generated with \code{\link{getMiRsites}}.}
}
\value{
A list.
}
\description{
The function rearrangeMiRres() rearranges the results of the
getMiRsites() function. Each element of the list contains the miR resutls
relative to one circRNA.
}
\examples{
# Load data frame containing predicted back-spliced junctions
data("mergedBSJunctions")

# Load short version of the gencode v19 annotation file
data("gtf")

# Annotate the first 3 back-spliced junctions
annotatedBSJs <- annotateBSJs(mergedBSJunctions[1:3, ], gtf,
    isRandom = FALSE)

# Retrieve target sequences.
 targets <- getCircSeqs(
     annotatedBSJs,
     gtf,
     species = "Hsapiens",
     genome = "hg19")

# Screen target sequence for miR binding sites.
pathToMiRs <- system.file("extdata", "miRs.txt",package="circRNAprofiler")

#miRsites <- getMiRsites(
#    targets,
#    species = "Hsapiens",
#    genome = "hg19",
#    miRspeciesCode = "hsa",
#    miRBaseLatestRelease = TRUE,
#    totalMatches = 6,
#    maxNonCanonicalMatches = 1,
#    pathToMiRs)

# Rearrange miR results
#rearragedMiRres <- rearrangeMiRres(miRsites)


}
