% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initCircRNAprofiler.R
\name{initCircRNAprofiler}
\alias{initCircRNAprofiler}
\title{Initialize the project folder}
\usage{
initCircRNAprofiler(projectFolderName, detectionTools)
}
\arguments{
\item{projectFolderName}{A character string specifying the name of the
project folder.}

\item{detectionTools}{A character vector specifying the tools
used to predict circRNAs. The following options are allowed: mapsplice,
nclscan, knife, circexplorer2, circmarker and uroborus. If the tool is
not mapsplice, nclscan, knife, circexplorer2, uroborus or circmarker
then use the option other. The user can choose 1 or multiple tools.
Subfolders named as the specified tools will be generated under
the working directory.}
}
\value{
A NULL object
}
\description{
The function initCircRNAprofiler() initializes the project
forlder.
}
\examples{
\dontrun{
initCircRNAprofiler(projectFolderName = "circProject",
    detectionTools = "mapsplice")
}


}
